requirements.txt
Faker==14.0.0
flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
flask_jwt_extended==4.4.4
Flask-RESTful==0.3.9
flask_sqlalchemy==3.0.2
wheel==0.37.0
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.2.2
python-dotenv==0.21.0
coverage==7.1.0
requests==2.26.0